pyproject.toml
[tool.poetry]
name = "textarria"
version = "0.1.0"
description = ""
authors = ["Bossguy124 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
mysql-connector = "^2.2.9"
tabulate = "^0.8.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
